Move windowed_query to separate file
